rename events and commands to old
